# Conflicts:
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/Sample.java
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/HelperClasses/Colors.java
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/OpModes/Camera/AutoTakeSample.java

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Intake/Extendo.java
@@ -90,7 +90,7 @@ public class Extendo {
         return getCurrentPosition() >= MaxExtendoExtension;
     }
 
-    public static double getMaxPosition() {
+    public static int getMaxPosition() {
         return MaxExtendoExtension;
     }
 }
